+++ mime-simple-events-calendar.php
@@ -3,7 +3,7 @@
  * Plugin Name:       MiMe Simple Events and Calendar
  * Plugin URI:        https://github.com/MilanMiMeOnline/WP-Simple-Events
  * Description:       A lightweight, native events plugin for WordPress.
- * Version:           0.2.3
+ * Version:           0.2.4
  * Requires at least: 6.9
  * Requires PHP:      8.3
  * Author:            MiMe
@@ -22,7 +22,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'WPSE_VERSION', '0.2.3' );
+define( 'WPSE_VERSION', '0.2.4' );
 define( 'WPSE_PLUGIN_FILE', __FILE__ );
 define( 'WPSE_PLUGIN_DIR', __DIR__ );